+++ pyccel/codegen/printing/cwrappercode.py
@@ -534,7 +534,10 @@ class CWrapperCodePrinter(CCodePrinter):
     def _print_Allocate(self, expr):
         variable = expr.variable
         if isinstance(variable.dtype, WrapperCustomDataType):
-            type_name = variable.cls_base.type_name
+            cls_base = variable.cls_base.original_class
+            class_def = self.scope.find(cls_base.scope.get_python_name(cls_base.name), 'classes')
+
+            type_name = class_def.type_name
             var_code = self._print(ObjectAddress(variable))
             decl_type = self.get_declare_type(variable)
             return f'{var_code} = ({decl_type}){type_name}.tp_alloc(&{type_name}, 0);\n'
@@ -544,7 +547,10 @@ class CWrapperCodePrinter(CCodePrinter):
     def _print_Deallocate(self, expr):
         variable = expr.variable
         if isinstance(variable.dtype, WrapperCustomDataType):
-            type_name = variable.cls_base.type_name
+            cls_base = variable.cls_base.original_class
+            class_def = self.scope.find(cls_base.scope.get_python_name(cls_base.name), 'classes')
+
+            type_name = class_def.type_name
             var_code = self._print(ObjectAddress(variable))
             return f'{type_name}.tp_free({var_code});\n'
         else: