+++ pyccel/parser/parser.py
@@ -105,6 +105,17 @@ class Parser(object):
         return self._filename
 
     @property
+    def original_filename(self) -> Path:
+        """
+        The absolute path to the original Python file that was translated.
+
+        This will be equivalent to the filename, unless the file is a dependency.
+        In that case the filename will be a .pyi file while the original_filename
+        will be a .py file.
+        """
+        return self._original_filename
+
+    @property
     def d_parsers(self):
         """Returns the d_parsers parser."""
 
@@ -288,7 +299,8 @@ class Parser(object):
             son = d_parsers_by_filename[str(filename)]
             son.append_parent(self)
             self.append_son(son)
-            d_parsers[source] = son
+            d_parsers[getattr(source, 'name', source)] = son
+            d_parsers.update(son.d_parsers)
 
         return d_parsers